Moved types and their methods in separate files to reduce clutter
